debt: refactor environment info collection and add helper file
